Known bugs: Edit cart - incorrect computation Edit product - not working Privacy - can be changed to About Us Cart count - can't display to all pages of customer Product Inventory - can't decrease product stocks when customer checks out Transaction Profile - displays all transactions even it's not from the account holder
Completed:
Registration Validations
Add To Cart QTY and Rent Duration Validations
Change Password Validations
Edit Cart - Save Edit Cart

+++ JewelryRentalSystem/Models/Cart.cs
@@ -15,7 +15,7 @@ namespace JewelryRentalSystem.Models
         public ApplicationUser Customer { get; set; } = null!;
 
         [DisplayName("Quantity")]
-        public double ProductQty { get; set; }
+        public int ProductQty { get; set; }
         [DisplayName("Rent Duration")]
         public int RentDuration { get; set; }
         public double Total { get; set; }
